Use ref struct in optimization algorithms (#7)
* use ref structure for optimization algorithms

+++ Arnible.MathModeling/Algebra/Polynomials/Polynomial.cs
@@ -148,7 +148,7 @@ namespace Arnible.MathModeling.Algebra.Polynomials
      * Query
      */
 
-    internal Dictionary<char, uint> GetIdentityVariableTerms()
+    internal Dictionary<char, ushort> GetIdentityVariableTerms()
     {
       return _terms.AsList().Select(t => t.GetIdentityVariableTerms())
         .AggregateCommonBy(v => v.Variable, vc => vc.Select(v => v.Power).MinDefensive());
